Fixed DDEX order book fetching when no orders on one of the sides
If no orders on one of the order book sides, `!xxx.length` was returning
true and the query always failed, even though the query could still be
fulfilled with orders on the other side of the orderbook.

At this moment, a good example is the `WBTC-WETH` book which has zero
asks. After this fix at least a WBTC sell query will be successful.