Unnecessary check for `totalPrice <= 0` removed
In case there is not enough liquidity, the `getPayAmount` and
`getBuyAmount` functions of the Eth2Dai contract just throw.
So if we get a result, it will always be a positive number.
No need to check it, it may actually mislead some future
developers which could think `0` is returned from the contract
in case there isn't enough liquidity.