Fixed support for RadarRelay markets which didn't fit on the first page
The first thing the RadarRelay client used to do was querying the list
of all markets (https://api.radarrelay.com/v2/markets) and looking for
our market in it. This approach wasn't giving correct results as the
markets list is paged, and most of markets do not fit on the first page
with the default `perPage=20`. Increasing `perPage` itself wasn't a
solution as the maximum value of this parameter in the RadarRelay API
is `100`, but the total number of markets is greater than that.

Instead of that I decided to query the individual market to verify its
existence. This way paging isn't an issue anymore.

With this fix, `WBTC-WETH` is now handled correctly.